Match container directory layout
Split source files into separate directories for app and MySQL
containers; not everything is needed by both.

Make the directory layout in git match the directory layout within the
containers so it's easier to understand what's used where.

Move Dockerfile into container-specific directory to prepare for the
future existence of multiple Dockerfiles, and to keep everything related
to each individual container in a container-specific directory.