Edit uid inside container to match host
Unlike some other container solutions, Docker doesn't (at least by
default?) configure uid mapping. This means that uids within the container
are identical to uids outside the container. In turn, this means that the
uid used to run daemons in the container must have access to the host
files located inside volume mounts. In general, the only uid that can
access files in a user's home directory is the uid of that user; there's
no guarantee that any other group, or everyone, will be able to access
them. Consequently, we must renumber uids inside the container to match
the owner of the files outside the container. This patch enhances the
entrypoint scripts to do just that.